refactor: organize rpc and api for server
